requirements.txt
absl-py==0.13.0
antlr4-python3-runtime==4.8
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.1
bleach==2.1.2
Brlapi==0.6.6
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
Cython==0.29.24
dataclasses==0.8
decorator==4.1.2
defer==1.0.6
devscripts===2.17.12ubuntu1.1
distro-info===0.18ubuntu0.18.04.1
docopt==0.6.2
entrypoints==0.2.3.post1
fairseq==0.10.2
filelock==3.0.12
Flask==2.0.1
gast==0.5.2
google-pasta==0.2.0
gpg==1.10.0
greenlet==0.4.12
grpcio==1.40.0
h5py==3.1.0
html5lib==0.999999999
httplib2==0.9.2
huggingface-hub==0.0.16
hydra-core==1.1.1
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.4
MarkupSafe==2.0.1
mistune==0.8.3
msgpack==0.5.6
nbconvert==5.3.1
nbformat==4.4.0
neovim==0.2.0
netifaces==0.10.4
nltk==3.6.2
notebook==5.2.2
num2words==0.5.10
numpy==1.19.5
oauth==1.0.1
olefile==0.45.1
omegaconf==2.1.1
packaging==21.0
pandas==1.1.5
pandocfilters==1.4.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.1.0
portalocker==2.3.2
prompt-toolkit==1.0.15
protobuf==3.17.3
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.5
python-crfsuite==0.9.7
python-dateutil==2.8.2
python-debian==0.1.32
python-magic==0.4.16
pytz==2021.1
pyxdg==0.25
PyYAML==5.4.1
pyzmq==16.0.2
regex==2021.8.28
reportlab==3.4.0
requests==2.26.0
requests-unixsocket==0.1.5
roman==3.3
sacrebleu==2.0.0
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.5.4
SecretStorage==2.3.1
sentencepiece==0.1.96
seqeval==1.2.2
simplegeneric==0.8.1
simplejson==3.13.2
six==1.16.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tabulate==0.8.9
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.7
testpath==0.3.1
threadpoolctl==2.2.0
tokenizers==0.10.3
torch==1.9.0
tornado==4.5.3
tqdm==4.62.2
traitlets==4.3.2
transformers==4.10.0
typing-extensions==3.10.0.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
underthesea==1.3.3
Unidecode==1.3.1
unidiff==0.5.4
urllib3==1.26.6
usb-creator==0.3.3
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5
Werkzeug==2.0.1
wrapt==1.12.1
xkit==0.0.0
zipp==3.5.0
zope.interface==4.3.2
